Performance boost: _.findWhere -> for loop
Replacing _.findWhere in _parseSample with a for loop results in a ~%50
speed-up loading a small VCF (as tested with Benchmark.js, and within
Chrome's profiler).

Also add in a little performance suite in test/perf.js (run with node
test/perf.js) in order to run little before/after tests for
statistically significant performance testing.